requirements.txt
torch
torchvision
transformers
streamlit
Pillow
numpy
scikit-learn
